requirements.txt
Flask==2.3.2
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-WTF==1.2.1
Jinja2==3.1.2
matplotlib==3.10.7
numpy==2.3.4
pycparser==2.21
pyparsing==3.2.5
python-dotenv==1.0.0
reportlab==4.4.5
requests==2.31.0